+++ test/unit/sessions/softphone-session-handler.test.ts
@@ -1,4 +1,4 @@
-import nock = require('nock');
+import nock from 'nock';
 
 import {
   SimpleMockSdk,
@@ -1415,6 +1415,16 @@ describe('getUserParticipantFromConversationEvent()', () => {
 
     expect(handler.getUserParticipantFromConversationEvent(converversationUpdate, CommunicationStates.alerting)).toEqual(participant2);
   });
+
+  it('should return the first participant that is not already terminated', () => {
+        /* setup our user to have multiple participants with different call states */
+        participant1.userId = userId;
+        participant1.calls = [{ ...call, state: CommunicationStates.terminated }];
+        participant2.userId = userId;
+        participant2.calls = [{ ...call }];
+        converversationUpdate.participants = [participant1, participant2];
+        expect(handler.getUserParticipantFromConversationEvent(converversationUpdate)).toEqual(participant2);
+  })
 
   it('should return the first participant with calls', () => {
     participant1.userId = userId;