+++ changelog.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.2.3...HEAD)
+### Fixed
+* [PCM-1911](https://inindca.atlassian.net/browse/PCM-1911) Added in logic to help with the ACD flow that checks for terminated/disconnected users if the user shows up twice in the participant update
 
 # [v7.2.3](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.2.2...v7.2.3)