+++ changelog.md
@@ -5,7 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.2.3...HEAD)
 ### Fixed
+<<<<<<< HEAD
 * [PCM-1911](https://inindca.atlassian.net/browse/PCM-1911) Added in logic to help with the ACD flow that checks for terminated/disconnected users if the user shows up twice in the participant update
+=======
+* [PCM-1913](https://inindca.atlassian.net/browse/PCM-1913) – Fixed the screen recording metadata request to use mids in place of trackIds.
+>>>>>>> 50747c5e3ac87f4fc2e49d0e6320f4ae4403a585
 
 # [v7.2.3](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.2.2...v7.2.3)