+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [PCM-1911](https://inindca.atlassian.net/browse/PCM-1911) Added in logic to help with the ACD flow that checks for terminated/disconnected users if the user shows up twice in the participant update
 * [PCM-1909](https://inindca.atlassian.net/browse/PCM-1909) – force terminate sessions so sdk.destroy doesn't hang in the case of persistent connection or LA==1
 * [PCM-1913](https://inindca.atlassian.net/browse/PCM-1913) – Fixed the screen recording metadata request to use mids in place of trackIds.
+* [PCM-1885](https://inindca.atlassian.net/browse/PCM-1885) – on `deviceschange` event from the window, SdkMedia should only validate all sessions if the `sessionManager` has already been initialized.
 
 # [v7.2.3](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.2.2...v7.2.3)